minimax

+++ src/state/state.hpp
@@ -8,7 +8,7 @@
 
 #include "../config.hpp"
 
-typedef std::pair<size_t, size_t> Point;
+typedef std::pair<int, int> Point;
 typedef std::pair<Point, Point> Move;
 class Board {
 public:
